remove api call for datacite and move archive method to publication manger
